feat: add PodPopup and convert ENABLE_RESOURCE_SCAN to bool
