+++ .env
@@ -45,4 +45,4 @@ USE_V2=
 ENABLE_EXTERNAL_ARGO_CD=false
 API_BATCH_SIZE=20
 SERVICE_WORKER_TIMEOUT='1'
-ENABLE_RESOURCE_SCAN="false"
+ENABLE_RESOURCE_SCAN=false
