+++ src/components/v2/appDetails/k8Resource/nodeType/types.ts
@@ -1,0 +1,10 @@
+import { NodeType } from '../../appDetails.type'
+import { NodeDetailTabsType } from '../../../../app/types'
+
+export interface PodPopupProps {
+    kind: NodeType
+    describeNode: (tab?: NodeDetailTabsType) => void
+    isExternalArgoApp: boolean
+    toggleShowDeleteConfirmation: () => void
+    handleShowVulnerabilityModal: () => void
+}
