+++ src/index.tsx
@@ -57,7 +57,7 @@ interface customEnv {
     ENABLE_EXTERNAL_ARGO_CD: boolean
     API_BATCH_SIZE: number
     SERVICE_WORKER_TIMEOUT?: string
-    ENABLE_RESOURCE_SCAN?: string
+    ENABLE_RESOURCE_SCAN?: boolean
 }
 declare global {
     interface Window {
@@ -176,7 +176,7 @@ if (!window || !window._env_) {
         ENABLE_EXTERNAL_ARGO_CD: false,
         API_BATCH_SIZE: 20,
         SERVICE_WORKER_TIMEOUT: '1',
-        ENABLE_RESOURCE_SCAN: 'false',
+        ENABLE_RESOURCE_SCAN: false,
     }
 }
